First cut of registering Python callbacks for xmlsec
This implementation uses a global linked-list structure to hold onto the
Python callbacks and registers a single C wrapper callback with xmlsec to
dispatch xmlsec's calling back to the appropriate Python function. This was
the simplest way I could think to emulate dynamically wrapping the Python
calls in C code, because C doesn't have closures.

A potential downside is that the state is all very global. Perhaps,
however, that's no worse than the very global set of callbacks that xmlsec
holds onto itself, so long as we don't have any strange threading stuff
going on.

In order to accommodate client code potentially interleaving registrations
of the default callbacks in between their own callbacks, the linked list
structure can have nodes with NULL function pointers in it, such that when
iterating through Python callbacks we can suspend our iteration and
delegate back to the defaults at the appropriate point.

As such, we always actually have the Python binding's C callbacks
registered with xmlsec, and every time we get asked to register the default
callbacks, we note a NULL function pointer and then re-register our C
callbacks.