Test and fix IO callback bindings
Turns out the `xmlSecAllIOCallbacks` pointer list yields the stored
callbacks in reverse, _and_ the default callbacks steal everything (
libxml2's `xmlFileMatch` is literally defined as `return(1)`! So we

 - Simplify our linked list of sets of Python callbacks to cons, rather
   than append
 - Don't bother trying to track interleaving default callbacks and Python
   callbacks, because `cur_cb_list_item` would get left in a bad state when
   the default matched.
 - Instead, drop all callbacks added prior to the default callbacks being
   added.

Tests-wise, there are a lot of failures relating to unreferenced objects
and/or memory leaks. However, as an example the top test is completely
commented out right now (i.e. doesn't do anything with the library), so
I'm a bit stuck as to how to grapple with that!