Don't define `MODULE_NAME` as a string literal
The `MODULE_NAME` macro is used in contexts where a string literal is
not valid, but the fallback value set in `src/common.h` defines it as
such; this differs from how it is defined in `setup.py`. Define
`MODULE_NAME` in `src/common.h` as it is defined in `setup.py`.

Fixes #267.